Protected pages created. Removed redundant code. Removed empty files for cleanup.

+++ next/hooks/useRemoveEvent.ts
@@ -1,14 +1,21 @@
 import apiClient from '@/services/apiClient';
+import { useState } from 'react';
 
 const useRemoveEvent = (
   eventId: string,
   communityId: string,
-  callback: () => void
+  callback: () => void,
+  setIsLoading: (value: boolean) => void
 ) => {
+
+  if (!eventId || !communityId) return;
+  setIsLoading(true);
+
   apiClient.delete(`/event/${eventId}`).then(() => {
-    apiClient.put(`/community/rmevent/${communityId}`, { eventId }).then(() => {
-      callback();
-    });
+    apiClient
+      .put(`/community/rmevent/${communityId}`, { eventId })
+      .then(() => callback())
+      .finally(() => setIsLoading(false));
   });
 };
